requirements.txt
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.9.0
argon2==0.1.10
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
certifi==2025.4.26
cffi==1.17.1
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.112.4
fastapi-cli==0.0.7
graphene==3.2.2
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
psycopg==3.2.9
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.6
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.38.6
starlette-graphene3==0.6.0
typer==0.15.4
typing-inspection==0.4.1
typing_extensions==4.13.2
uvicorn==0.22.0
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
